+++ src/Check/Redis/RedisExt/RedisPingPongCheck.php
@@ -53,7 +53,7 @@ readonly class RedisPingPongCheck implements CheckInterface
 
     public function getExtraParameters(): array
     {
-        // By security we not return password (because many redis instances work in internal network).
+        // By security, we not return password (because many redis instances work in internal network).
         return [
             'host' => $this->host,
             'port' => $this->port,