+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 CHANGELOG
 =========
 
-Next release
-------------
+v2.1.1
+------
 
 Add next checks: